Update project gradle settings for latest gradle

+++ app/src/main/java/com/sahil_madan/guitarprotabs/GPTab.java
@@ -6,11 +6,11 @@ package com.sahil_madan.guitarprotabs;
 public class GPTab {
     public String artist;
     public String song;
-    public int rating;
+    public double rating;
     public int votes;
     public String link;
 
-    public GPTab(String artist, String song, int rating, int votes, String link) {
+    public GPTab(String artist, String song, double rating, int votes, String link) {
         this.artist = artist;
         this.song = song;
         this.rating = rating;
